fix: pre-commit hookによるlinter修正とアーカイブファイル追加
- Black, isortによるコード整形
- archiveディレクトリのバックアップファイル追加
